+++ code/include/ACSException.hpp
@@ -3,8 +3,8 @@
  * @brief Exception class for handling ACS-related errors
  *
  * @author Francesco Biscaccia Carrara
- * @version v1.2.9
- * @since 07/06/2025
+ * @version v1.2.10
+ * @since 07/09/2025
  */
 
 #ifndef ACS_EXC_H