+++ code/include/FMIP.hpp
@@ -7,8 +7,8 @@
  * It inherits core functionality from the base MIP class while adding FMIP-specific features.
  *
  * @author Francesco Biscaccia Carrara
- * @version v1.2.9
- * @since 07/06/2025
+ * @version v1.2.10
+ * @since 07/09/2025
  */
 
 #ifndef FMIP_H