+++ Changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.2.10] - 2025-07-09  
+### Fixed  
+- Optimized `FixPolicy::walkMIPMT` by reducing redundant computations through in-place recomputation of violated constraints during execution ([FixPolicy.cpp](code/source/FixPolicy.cpp)).  
+- Improved performance of `MIP::getViolatedConstrIndex` by operating on references instead of returning a `std::vector<int>` ([MIP.hpp](code/include/MIP.hpp)).  
+
+
 ## [1.2.9] - 2025-07-06  
 ### Fixed  
 - Optimized `FixPolicy::walkMIPMT` to reduce computational overhead while enhancing solution diversification ([FixPolicy.cpp](code/source/FixPolicy.cpp)).